rename events to whats on, improve structure
