+++ packages/website/src/pages/whatson/branding/map.tsx
@@ -33,7 +33,7 @@ export const whatsOnBrandingMap: {
     [WhatsOnBrandedComponentLocation.Container]: ({ children }) => (
       <div
         css={{
-          backgroundImage: `url(${require('../../../../img/freshers-2019/repeatable-bg-left-01.svg')}), url(${require('../../../../img/freshers-2019/repeatable-bg-right-02.svg')})`,
+          backgroundImage: `url(${require('../../../img/freshers-2019/repeatable-bg-left-01.svg')}), url(${require('../../../img/freshers-2019/repeatable-bg-right-02.svg')})`,
           backgroundRepeat: 'repeat-y, repeat-y',
           backgroundPosition: 'top left, top right',
           backgroundSize: '20%',
